requirements.txt
facenet_pytorch==2.5.3
numpy==1.26.4
Pillow==10.2.0
scikit_learn==1.4.1.post1
torch==2.2.0
torchvision==0.17.0
